Created form component for todo form
