+++ src/components/Form.jsx
@@ -1,0 +1,23 @@
+import { useState } from "react";
+
+export default function Form({todos={todos} ,setTodos={setTodos}}) {
+  const [todo, setTodo] = useState("");
+
+  function handleSubmit(e) {
+    e.preventDefault();
+    setTodos([...todos, todo]);
+    setTodo("");
+  }
+  return (
+    <div>
+      <form onSubmit={handleSubmit}>
+        <input
+          onChange={(e) => setTodo(e.target.value)}
+          value={todo}
+          type="text"
+        />
+        <button type="submit"> Add </button>
+      </form>
+    </div>
+  );
+}
