+++ src/components/Todo.jsx
@@ -1,24 +1,12 @@
 import { useState } from "react";
 import TodoItem from "./TodoItem";
+import Form from "./Form";
 
 export default function Todo() {
-  const [todo, setTodo] = useState("");
   const [todos, setTodos] = useState([]);
-  function handleSubmit(e) {
-    e.preventDefault();
-    setTodos([...todos, todo]);
-    setTodo("");
-  }
   return (
     <div>
-      <form onSubmit={handleSubmit}>
-        <input
-          onChange={(e) => setTodo(e.target.value)}
-          value={todo}
-          type="text"
-        />
-        <button type="submit"> Add </button>
-      </form>
+      <Form  todos={todos} setTodos={setTodos}/>
       {todos.map((item) => (
         <TodoItem key={item} item={item} />
       ))}
